[#1596] move task to documenten app
